Address review: rename to ExplicitAttributeNamedArgsRector, make non-configurable (name all attributes), move to named-args set
