--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/AttributeNamedArgsRectorFirstNamedPositionTest.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector;
-
-use Iterator;
-use PHPUnit\Framework\Attributes\DataProvider;
-use Rector\Testing\PHPUnit\AbstractRectorTestCase;
-
-final class AttributeNamedArgsRectorFirstNamedPositionTest extends AbstractRectorTestCase
-{
-    #[DataProvider('provideData')]
-    public function test(string $filePath): void
-    {
-        $this->doTestFile($filePath);
-    }
-
-    public static function provideData(): Iterator
-    {
-        return self::yieldFilesFromDirectory(__DIR__ . '/FixtureFirstNamedPosition');
-    }
-
-    public function provideConfigFilePath(): string
-    {
-        return __DIR__ . '/config/configured_rule_first_named_position.php';
-    }
-}
