--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/Fixture/positional_middleware.php.inc
@@ -1,27 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Fixture;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[Middleware(RequireToken::class, except: ['home'])]
-class PositionalMiddleware
-{
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Fixture;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[Middleware(middleware: RequireToken::class, except: ['home'])]
-class PositionalMiddleware
-{
-}
-
-?>
