--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/Fixture/skip_already_named.php.inc
@@ -1,11 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Fixture;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[Middleware(middleware: RequireToken::class, only: ['export'])]
-class SkipAlreadyNamed
-{
-}
