--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/Fixture/skip_other_attribute.php.inc
@@ -1,10 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Fixture;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\OtherAttribute;
-
-#[OtherAttribute('a', 'b')]
-class SkipOtherAttribute
-{
-}
