--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/Fixture/variadic_named_prefix.php.inc
@@ -1,27 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Fixture;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\MiddlewareWithVariadic;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[MiddlewareWithVariadic(RequireToken::class)]
-class VariadicNamedPrefix
-{
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Fixture;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\MiddlewareWithVariadic;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[MiddlewareWithVariadic(middleware: RequireToken::class)]
-class VariadicNamedPrefix
-{
-}
-
-?>
