--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/FixtureFirstNamedPosition/positional_except.php.inc
@@ -1,27 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\FixtureFirstNamedPosition;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[Middleware(RequireToken::class, null, ['home'])]
-class PositionalExcept
-{
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\FixtureFirstNamedPosition;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[Middleware(RequireToken::class, only: null, except: ['home'])]
-class PositionalExcept
-{
-}
-
-?>
