--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/FixtureFirstNamedPosition/skip_middleware_left_positional.php.inc
@@ -1,11 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\FixtureFirstNamedPosition;
-
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\RequireToken;
-
-#[Middleware(RequireToken::class, except: ['home'])]
-class SkipMiddlewareLeftPositional
-{
-}
