--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/Source/RequireToken.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source;
-
-class RequireToken
-{
-}
