--- rules-tests/CodeQuality/Rector/Attribute/AttributeNamedArgsRector/config/configured_rule_first_named_position.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\CodeQuality\Rector\Attribute\AttributeNamedArgsRector;
-use Rector\CodeQuality\ValueObject\AttributeNamedArgs;
-use Rector\Config\RectorConfig;
-use Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source\Middleware;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->ruleWithConfiguration(AttributeNamedArgsRector::class, [
-        new AttributeNamedArgs(Middleware::class, 1),
-    ]);
-};
