+++ rules-tests/CodeQuality/Rector/Attribute/ExplicitAttributeNamedArgsRector/Fixture/other_attribute.php.inc
@@ -1,0 +1,25 @@
+<?php
+
+namespace Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Fixture;
+
+use Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Source\OtherAttribute;
+
+#[OtherAttribute('a', 'b')]
+class OtherAttributeFixture
+{
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Fixture;
+
+use Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Source\OtherAttribute;
+
+#[OtherAttribute(a: 'a', b: 'b')]
+class OtherAttributeFixture
+{
+}
+
+?>
