+++ rules-tests/CodeQuality/Rector/Attribute/ExplicitAttributeNamedArgsRector/Fixture/skip_already_named.php.inc
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Fixture;
+
+use Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Source\Middleware;
+use Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Source\RequireToken;
+
+#[Middleware(middleware: RequireToken::class, only: ['export'])]
+class SkipAlreadyNamed
+{
+}
