+++ rules-tests/CodeQuality/Rector/Attribute/ExplicitAttributeNamedArgsRector/Source/MiddlewareWithVariadic.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\CodeQuality\Rector\Attribute\AttributeNamedArgsRector\Source;
+namespace Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Source;
 
 #[\Attribute]
 class MiddlewareWithVariadic
