+++ rules-tests/CodeQuality/Rector/Attribute/ExplicitAttributeNamedArgsRector/Source/RequireToken.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rector\Tests\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector\Source;
+
+class RequireToken
+{
+}
