+++ rules-tests/CodeQuality/Rector/Attribute/ExplicitAttributeNamedArgsRector/config/configured_rule.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\CodeQuality\Rector\Attribute\ExplicitAttributeNamedArgsRector;
+use Rector\Config\RectorConfig;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->rule(ExplicitAttributeNamedArgsRector::class);
+};
